ajout de page niveau && classe almost there :(

+++ database/migrations/2023_03_24_112114_inscriptions.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->date('date_inscription')->default(now());
             $table->year('annee_scolaire')->default(now()->format('Y'));
-            $table->integer('id_niveaux');
+            $table->unsignedBigInteger('id_niveaux');
             $table->foreign('id_niveaux')->references('id')->on('niveaux')->onDelete('cascade');
             $table->unsignedBigInteger('id_filieres');
             $table->foreign('id_filieres')->references('id')->on('filieres')->onDelete('cascade');
