+++ database/migrations/2023_03_28_095259_niveaux_filieres.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('niveaux_filieres',function(Blueprint $table){
-            $table->integer('id_niveaux');
+            $table->unsignedBigInteger('id_niveaux');
             $table->unsignedBigInteger('id_filieres');
             $table->foreign('id_niveaux')->references('id')->on('niveaux');
             $table->foreign('id_filieres')->references('id')->on('filieres');
